% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcasHLATools.R
\docType{package}
\name{arcasHLATools-package}
\alias{arcasHLATools-package}
\alias{arcasHLATools}
\title{arcasHLATools: TODO}
\description{
This package TODO.
}
\seealso{
\itemize{
\item \code{\link{readAllArcasHLAGenotypefromDir}} {for extracting typing
from multiple arcasHLA gentype files present in one directory. }
\item \code{\link{readOneArcasHLAGenotypeFile}} { for extracting typing from
one arcasHLA genotype file. }
}
}
\author{
Astrid Deschenes and
Pascal Belleau

Maintainer:
Astrid Deschenes <adeschen@hotmail.com>
}
\keyword{package}
