% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{readAllArcasHLAGenotypefromDir}
\alias{readAllArcasHLAGenotypefromDir}
\title{Extract all HLA typing from all genotypes json files present
in a directory}
\usage{
readAllArcasHLAGenotypefromDir(directory_path, prefix = "",
  suffix = ".genotype.json", recursive = FALSE)
}
\arguments{
\item{directory_path}{a \code{character} string corresponding to the
name of the directory containing the arcasHLA
genotype files.}

\item{prefix}{a \code{character} string specifying the prefix of the files
to load. It is usefull when only a subsection of the files need
to be loaded. Default = \code{""}.}

\item{suffix}{a \code{character} string specifying the suffix of the files
to load. It is usefull when only a subsection of the files need
to be loaded. Default = \code{".genotype.json"}.}

\item{recursive}{a \code{logical} indicating if the listing should
recurse into subdirectories. Default = \code{FALSE}.}
}
\value{
a TODO
}
\description{
TODO
}
\examples{

## Directory containing arcasHLA genotype files
directory <- system.file("extdata/arcasHLAFiles", package = "arcasHLATools")

## Reading all genotype files in the directory
readAllArcasHLAGenotypefromDir(directory)

}
\author{
Astrid Deschenes
}
