% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{readOneArcasHLAGenotypeFile}
\alias{readOneArcasHLAGenotypeFile}
\title{Extract all HLA typing from one genotypes json file}
\usage{
readOneArcasHLAGenotypeFile(file_path, extension = ".genotype.json")
}
\arguments{
\item{file_path}{a \code{character} string corresponding to the
name of the arcasHLA genotype file.}

\item{extension}{a \code{character} string correpondint to the extension
of the genotype file. The extension will be removed from the file name which
is going to be assigned to the dataset. Default=".genotype.json".}
}
\value{
TODO
}
\description{
TODO
}
\examples{


## Path to an arcasHLA genotype file
jsonFile <- system.file("extdata/arcasHLAFiles", "Sample002.genotype.json",
    package = "arcasHLATools")

## Extracting file information
readOneArcasHLAGenotypeFile(jsonFile)

}
\author{
Astrid Deschenes
}
